Modify dashboard to fetch repo details with graph
